+++ src/hooks/useCalendarRange.test.tsx
@@ -1,6 +1,7 @@
-import dayjs from 'dayjs';
+import React from 'react';
 import { dateTime, TimeRange } from '@grafana/data';
-import { act, renderHook } from '@testing-library/react';
+import { act, render, renderHook, screen } from '@testing-library/react';
+import dayjs from 'dayjs';
 import { View } from '../types';
 import { getUnitType, useCalendarRange } from './useCalendarRange';
 
@@ -31,15 +32,66 @@ describe('Use Calendar Range', () => {
     onChangeTimeRange.mockClear();
   });
 
-  it('Should set last month for initial date', () => {
+  it('Should set last month for initial date if month by default', () => {
     const last3Months = dayjs(getSafeDate()).subtract(3, 'month');
     const timeRange = {
       ...defaultTimeRange,
       from: dateTime(last3Months.toDate()),
     };
-    const { result } = renderHook(() => useCalendarRange(timeRange, onChangeTimeRange));
+    const { result } = renderHook(() => useCalendarRange(timeRange, onChangeTimeRange, View.MONTH));
 
     expect(result.current.date.toISOString()).toEqual(new Date('2023-02-01 12:00').toISOString());
+  });
+
+  it('Should set last week for initial date if week by default', () => {
+    const last3Months = dayjs(getSafeDate()).subtract(3, 'month');
+    const timeRange = {
+      ...defaultTimeRange,
+      from: dateTime(last3Months.toDate()),
+    };
+    const { result } = renderHook(() => useCalendarRange(timeRange, onChangeTimeRange, View.WEEK));
+
+    expect(result.current.date.toISOString()).toEqual(new Date('2023-01-31 00:00').toISOString());
+  });
+
+  it('Should set last day for initial date if day by default', () => {
+    const last3Months = dayjs(getSafeDate()).subtract(3, 'month');
+    const timeRange = {
+      ...defaultTimeRange,
+      from: dateTime(last3Months.toDate()),
+    };
+    const { result } = renderHook(() => useCalendarRange(timeRange, onChangeTimeRange, View.DAY));
+
+    expect(result.current.date.toISOString()).toEqual(new Date('2023-02-02 00:00').toISOString());
+  });
+
+  it('Should update calendar dates on time range update', () => {
+    const last3Months = dayjs(getSafeDate()).subtract(3, 'month');
+    const timeRange = {
+      ...defaultTimeRange,
+      from: dateTime(last3Months.toDate()),
+    };
+
+    /**
+     * Component to test re-render
+     * @param timeRange
+     * @constructor
+     */
+    const Component = ({ timeRange }: any) => {
+      const { date } = useCalendarRange(timeRange, onChangeTimeRange, View.DAY);
+      return <div data-testid="date">{date.toISOString()}</div>;
+    };
+
+    const { rerender } = render(<Component timeRange={timeRange} />);
+
+    expect(screen.getByTestId('date')).toHaveTextContent('2023-02-02T00:00:00.000Z');
+
+    /**
+     * Re-render with updated timeRange
+     */
+    rerender(<Component timeRange={{ ...timeRange, from: dateTime(last3Months.subtract(1, 'month').toDate()) }} />);
+
+    expect(screen.getByTestId('date')).toHaveTextContent('2022-12-02T12:00:00.000Z');
   });
 
   describe('Navigate', () => {
@@ -49,7 +101,7 @@ describe('Use Calendar Range', () => {
         ...defaultTimeRange,
         from: dateTime(last3Months.toDate()),
       };
-      const { result } = renderHook(() => useCalendarRange(timeRange, onChangeTimeRange));
+      const { result } = renderHook(() => useCalendarRange(timeRange, onChangeTimeRange, View.MONTH));
 
       /**
        * Month